[refact] Renamed files and code variables for PR comments.
