+++ transfer_queue/storage/__init__.py
@@ -13,7 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .backends import SimpleStorageUnit, StorageUnitData
 from .managers import (
     AsyncSimpleStorageManager,
     MooncakeStorageManager,
@@ -22,6 +21,7 @@ from .managers import (
     StorageManagerFactory,
     YuanrongStorageManager,
 )
+from .simple_storage import SimpleStorageUnit, StorageUnitData
 
 __all__ = [
     "SimpleStorageUnit",
